requirements.txt
chainlit
notebook
ipywidgets
tqdm
python-dotenv
langchain-google-genai

-e .